Updated performance test suite. Modified json files, moved channel management endpoints under data management, added new simulation 'get payload'
