+++ src/gatling/resources/data/daily-cause-list-civil-and-family/civilAndFamilyDailyCauseList.json
@@ -29,7 +29,6 @@
             "courtRoomName": "1",
             "session": [
               {
-                "sessionStartTime": "",
                 "sessionChannel": ["VIDEO HEARING"],
                 "judiciary": [
                   {
@@ -181,7 +180,6 @@
             "courtRoomName": "1",
             "session": [
               {
-                "sessionStartTime": "",
                 "sessionChannel": ["VIDEO HEARING"],
                 "judiciary": [
                   {
