+++ src/gatling/resources/data/daily-cause-list-civil-and-family/civilAndFamilyDailyCauseList2MB.json
@@ -35,7 +35,6 @@
             "courtRoomName": "1",
             "session": [
               {
-                "sessionStartTime": "",
                 "sessionChannel": [
                   "VIDEO HEARING"
                 ],
