+++ src/gatling/resources/data/daily-cause-list-civil/civilDailyCauseList.json
@@ -30,7 +30,6 @@
             "courtRoomName": "1",
             "session": [
               {
-                "sessionStartTime": "",
                 "sessionChannel": ["VIDEO HEARING"],
                 "judiciary": [
                   {
@@ -122,7 +121,6 @@
             "courtRoomName": "1",
             "session": [
               {
-                "sessionStartTime": "",
                 "sessionChannel": ["VIDEO HEARING"],
                 "judiciary": [
                   {
@@ -170,7 +168,6 @@
             "courtRoomName": "1",
             "session": [
               {
-                "sessionStartTime": "",
                 "sessionChannel": ["VIDEO HEARING"],
                 "judiciary": [
                   {
@@ -218,7 +215,6 @@
             "courtRoomName": "1",
             "session": [
               {
-                "sessionStartTime": "",
                 "sessionChannel": ["VIDEO HEARING"],
                 "judiciary": [
                   {
