+++ src/gatling/resources/data/daily-cause-list-family/familyDailyCauseList.json
@@ -29,7 +29,6 @@
             "courtRoomName": "1",
             "session": [
               {
-                "sessionStartTime": "",
                 "sessionChannel": ["VIDEO HEARING"],
                 "judiciary": [
                   {
@@ -257,7 +256,6 @@
             "courtRoomName": "1",
             "session": [
               {
-                "sessionStartTime": "",
                 "sessionChannel": ["VIDEO HEARING"],
                 "judiciary": [
                   {
